Move `HistoryFileFormat` to history
