+++ crates/nu-protocol/src/config/reedline.rs
@@ -59,26 +59,6 @@ pub struct CursorShapeConfig {
     pub vi_normal: NuCursorShape,
 }
 
-#[derive(Clone, Copy, Debug, IntoValue, PartialEq, Eq, Serialize, Deserialize)]
-pub enum HistoryFileFormat {
-    /// Store history as an SQLite database with additional context
-    Sqlite,
-    /// store history as a plain text file where every line is one command (without any context such as timestamps)
-    Plaintext,
-}
-
-impl FromStr for HistoryFileFormat {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_ascii_lowercase().as_str() {
-            "sqlite" => Ok(Self::Sqlite),
-            "plaintext" => Ok(Self::Plaintext),
-            _ => Err("expected either 'sqlite' or 'plaintext'"),
-        }
-    }
-}
-
 #[derive(Clone, Copy, Debug, Default, IntoValue, PartialEq, Eq, Serialize, Deserialize)]
 pub enum EditBindings {
     Vi,
